+++ mango3-core/src/models/mod.rs
@@ -1,16 +1,20 @@
 use argon2::password_hash::rand_core::OsRng;
 use argon2::password_hash::SaltString;
 use argon2::{Argon2, PasswordHash, PasswordHasher, PasswordVerifier};
+use rand::distributions::Alphanumeric;
+use rand::{thread_rng, Rng};
 use rust_iso3166::CountryCode;
 use sqlx::types::chrono::NaiveDate;
 use sqlx::types::Uuid;
 
 mod blob;
+mod confirmation_code;
 mod user;
 mod user_session;
 mod website;
 
 pub use blob::Blob;
+pub use confirmation_code::ConfirmationCode;
 pub use user::User;
 pub use user_session::UserSession;
 pub use website::Website;
@@ -23,6 +27,14 @@ fn encrypt_password(password: &str) -> String {
 
 fn find_country(query: &str) -> Option<&CountryCode> {
     rust_iso3166::ALL.iter().find(|c| c.alpha2 == query || c.name == query)
+}
+
+pub fn generate_random_string(length: i8) -> String {
+    thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(length as usize)
+        .map(char::from)
+        .collect()
 }
 
 fn parse_date(value: &str) -> Option<NaiveDate> {