+++ mango3-leptos-utils/src/components/action_form_alert.rs
@@ -6,40 +6,55 @@ use server_fn::error::NoCustomError;
 
 use crate::models::ActionFormResp;
 
+use super::BoxedFn;
+
 #[component]
 pub fn ActionFormAlert(
     action_value: RwSignal<Option<Result<ActionFormResp, ServerFnError<NoCustomError>>>>,
     #[prop(into)] error_message: TextProp,
-    #[prop(into)] redirect_to: String,
+    #[prop(into, optional)] on_success: Option<BoxedFn>,
+    #[prop(into, optional)] redirect_to: Option<String>,
     #[prop(into)] success_message: TextProp,
 ) -> impl IntoView {
     let navigate = use_navigate();
+    let is_done = RwSignal::new(false);
+
+    Effect::new(move || {
+        if !is_done.get() {
+            return;
+        }
+
+        if let Some(on_success) = on_success.as_ref() {
+            on_success.0();
+        }
+
+        if let Some(to) = redirect_to.clone() {
+            navigate(&to, Default::default())
+        }
+    });
+
     let error_msg = move || error_message.clone().get();
     let success_msg = move || success_message.clone().get();
 
     move || match ActionFormResp::from(action_value).success {
-        Some(true) => {
-            let navigate = navigate.clone();
-            let redirect_to = redirect_to.clone();
-            EitherOf3::A(view! {
-                <dialog class="modal modal-open">
-                    <div class="modal-box">
-                        <div>{success_msg.clone()}</div>
-                        <div class="modal-action">
-                            <button
-                                class="btn btn-primary"
-                                on:click=move |event| {
-                                    event.prevent_default();
-                                    navigate(&redirect_to.clone(), Default::default())
-                                }
-                            >
-                                "Ok"
-                            </button>
-                        </div>
+        Some(true) => EitherOf3::A(view! {
+            <dialog class="modal" class:modal-open=move || !is_done.get()>
+                <div class="modal-box">
+                    <div>{success_msg.clone()}</div>
+                    <div class="modal-action">
+                        <button
+                            class="btn btn-primary"
+                            on:click=move |event| {
+                                event.prevent_default();
+                                is_done.set(true);
+                            }
+                        >
+                            "Ok"
+                        </button>
                     </div>
-                </dialog>
-            })
-        }
+                </div>
+            </dialog>
+        }),
         Some(false) => EitherOf3::B(view! {
             <div class="pt-2 pb-2">
                 <div role="alert" class="alert alert-error">