+++ mango3-leptos-utils/src/models/user_resp.rs
@@ -13,6 +13,8 @@ pub struct UserResp {
     pub username: String,
     pub display_name: String,
     pub initials: String,
+    pub email: String,
+    pub email_is_confirmed: bool,
     pub avatar_image_blob: Option<BlobResp>,
 }
 
@@ -24,6 +26,8 @@ impl UserResp {
             username: user.username.clone(),
             display_name: user.display_name.clone(),
             initials: user.initials(),
+            email: user.email.clone(),
+            email_is_confirmed: user.email_is_confirmed(),
             avatar_image_blob: user
                 .avatar_image_blob(&core_context)
                 .await