+++ mango3-monitor/src/constants.rs
@@ -1,4 +1,7 @@
+pub const KEY_TEXT_CONFIRMATION_CODE: &str = "confirmation-code";
 pub const KEY_TEXT_HELLO: &str = "hello";
+pub const KEY_TEXT_USE_THIS_CODE_TO_ACTION: &str = "use-this-code-to-action";
 pub const KEY_TEXT_WELCOME_TO_TITLE: &str = "welcome-to-title";
 
+pub const KEY_TEXT_ARG_ACTION: &str = "action";
 pub const KEY_TEXT_ARG_TITLE: &str = "title";