+++ mango3-my-account/src/pages/edit_email_page.rs
@@ -1,0 +1,87 @@
+use leptos::prelude::*;
+use leptos_i18n::t_string;
+
+use mango3_leptos_utils::components::{ActionFormAlert, CurrentUserResource, PasswordField, SubmitButton, TextField};
+use mango3_leptos_utils::context::use_current_user_resource;
+use mango3_leptos_utils::i18n::{t, use_i18n};
+use mango3_leptos_utils::models::ActionFormResp;
+use mango3_leptos_utils::pages::AuthenticatedPage;
+
+use crate::components::EmailConfirmationBadge;
+use crate::server_functions::AttemptToUpdateEmail;
+
+#[component]
+pub fn EditEmailPage() -> impl IntoView {
+    let i18n = use_i18n();
+    let current_user_resource = use_current_user_resource();
+    let server_action = ServerAction::<AttemptToUpdateEmail>::new();
+    let action_value = server_action.value();
+    let error_email = RwSignal::new(None);
+    let error_password = RwSignal::new(None);
+
+    Effect::new(move || {
+        let response = ActionFormResp::from(action_value);
+
+        error_email.set(response.error("email"));
+        error_password.set(response.error("password"));
+    });
+
+    let title = move || t_string!(i18n, my_account.edit_email);
+
+    view! {
+        <AuthenticatedPage title=title>
+            <h2 class="h2">{title}</h2>
+
+            <section class="max-w-[640px] w-full ml-auto mr-auto">
+                <h3 class="text-lg font-bold mb-4">{t!(i18n, my_account.current_email)}</h3>
+
+                <div class="flex items-center justify-between">
+                    <CurrentUserResource children=move |user| {
+                        user.map(|user| {
+                            let email_is_confirmed = RwSignal::new(user.email_is_confirmed);
+                            view! {
+                                <span>{user.email}</span>
+
+                                <EmailConfirmationBadge is_confirmed=email_is_confirmed />
+                            }
+                        })
+                    } />
+                </div>
+            </section>
+
+            <section class="max-w-[640px] w-full ml-auto mr-auto mt-4">
+                <h3 class="text-lg font-bold mb-4">{t!(i18n, my_account.change_email)}</h3>
+
+                <ActionForm
+                    action=server_action
+                    attr:autocomplete="off"
+                    attr:novalidate="true"
+                    attr:class="form"
+                >
+                    <ActionFormAlert
+                        action_value=action_value
+                        error_message=move || t_string!(i18n, my_account.failed_to_update_email)
+                        on_success=move || current_user_resource.refetch()
+                        success_message=move || {
+                            t_string!(i18n, my_account.email_updated_successfully)
+                        }
+                    />
+
+                    <TextField
+                        label=move || t_string!(i18n, shared.email)
+                        name="email"
+                        error=error_email
+                    />
+
+                    <PasswordField
+                        label=move || t_string!(i18n, shared.password)
+                        name="password"
+                        error=error_password
+                    />
+
+                    <SubmitButton is_loading=server_action.pending() />
+                </ActionForm>
+            </section>
+        </AuthenticatedPage>
+    }
+}