+++ mango3-my-account/src/pages/mod.rs
@@ -1,9 +1,11 @@
 mod change_password_page;
+mod edit_email_page;
 mod edit_profile_page;
 mod index_page;
 mod index_parent_page;
 
 pub use change_password_page::ChangePasswordPage;
+pub use edit_email_page::EditEmailPage;
 pub use edit_profile_page::EditProfilePage;
 pub use index_page::IndexPage;
 pub use index_parent_page::IndexParentPage;