+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Features to update and show websites.
 - Feature to update profile.
 - Features to show websites list.
+- Feature to confirm email.